requirements.txt
datasets
transformers
monai
torch
torchvision
tqdm
numpy
pillow
tifffile[all]
pyometiff